Extracting #share_notified? behavior
This pull request provides more explicit logic guarding the interface
expectations of the `#share_notified?` method. In other words, by
leveraging a collaborating class, we can instantiate that class and
enforce the expected behavior.

Also, instead of mixing in multiple methods, in particular one that was
named `#call`, I am preferring to instead mix in a smaller number of
methods.